Extract request convenience methods from Checkout VC
